+++ src/main/java/at/hannibal2/skyhanni/events/entity/EntityMaxHealthUpdateEvent.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.events
+package at.hannibal2.skyhanni.events.entity
 
 import at.hannibal2.skyhanni.api.event.SkyHanniEvent
 import net.minecraft.entity.EntityLivingBase
