+++ src/main/java/at/hannibal2/skyhanni/events/garden/farming/FarmingContestEvent.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.events
+package at.hannibal2.skyhanni.events.garden.farming
 
 import at.hannibal2.skyhanni.api.event.SkyHanniEvent
 import at.hannibal2.skyhanni.features.garden.CropType
