+++ src/main/java/at/hannibal2/skyhanni/mixins/transformers/MixinNetHandlerPlayClient.java
@@ -1,6 +1,6 @@
 package at.hannibal2.skyhanni.mixins.transformers;
 
-import at.hannibal2.skyhanni.events.EntityEquipmentChangeEvent;
+import at.hannibal2.skyhanni.events.entity.EntityEquipmentChangeEvent;
 import at.hannibal2.skyhanni.events.minecraft.packet.PacketSentEvent;
 import net.minecraft.client.network.NetHandlerPlayClient;
 import net.minecraft.entity.Entity;
